add some initial docs
