+++ app/assets/stylesheets/components/autocomplete.scss
@@ -2,7 +2,7 @@
 
 .crayons-autocomplete {
   &__popover {
-    min-width: 250px;
+    min-width: 300px;
     padding: var(--su-1);
 
     &[data-reach-popover] {
