+++ app/javascript/crayons/MentionAutocomplete/__stories__/MentionAutocomplete.stories.jsx
@@ -5,9 +5,32 @@ export default {
   title: 'Components/MentionAutocomplete',
 };
 
+function fetchUsers(searchTerm) {
+  const exampleApiResult = {
+    result: [
+      {
+        username: 'one',
+        name: 'First name Last Name One Two Three',
+        profile_image_90: 'https://placedog.net/50',
+      },
+      {
+        username: 'two',
+        name: 'User Two',
+        profile_image_90: 'https://placedog.net/51',
+      },
+    ],
+  };
+
+  return Promise.resolve(
+    exampleApiResult.result.filter((user) =>
+      user.username.includes(searchTerm),
+    ),
+  );
+}
+
 export const Default = () => (
   <div>
-    <MentionAutocomplete onSelect={() => console.log('selected')} />
+    <MentionAutocomplete onSelect={() => {}} fetchSuggestions={fetchUsers} />
   </div>
 );
