+++ app/javascript/crayons/MentionAutocomplete/__stories__/mention-autocomplete.md
@@ -1,0 +1,7 @@
+## Mention Autocomplete
+
+** This component is a work in progress **
+
+The mention autocomplete component uses [Reach UI's Combobox](https://reach.tech/combobox/) under the hood. It can be used for searching for users by username by passing in the `fetchSuggestions` prop.
+
+As a general rule, no more than 6 suggestions should be shown at a time.
